Move odds calculator types to triplea.odds.calculator package
This fixes several violations of the Checkstyle PackageName rule.

BattleResults cannot be moved at this time because instances of this
class may be stored in a save game.  Changing the fully-qualified class
name will break compatibility.
