+++ game-core/src/main/java/games/strategy/triplea/ai/fast/AggregateEstimate.java
@@ -7,7 +7,7 @@ import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.PlayerID;
 import games.strategy.engine.data.Unit;
 import games.strategy.engine.data.UnitType;
-import games.strategy.triplea.oddsCalculator.ta.AggregateResults;
+import games.strategy.triplea.odds.calculator.AggregateResults;
 import games.strategy.triplea.util.TuvUtils;
 import games.strategy.util.IntegerMap;
 import games.strategy.util.Tuple;
