+++ game-core/src/main/java/games/strategy/triplea/ai/pro/ProAi.java
@@ -45,8 +45,8 @@ import games.strategy.triplea.delegate.remote.IAbstractPlaceDelegate;
 import games.strategy.triplea.delegate.remote.IMoveDelegate;
 import games.strategy.triplea.delegate.remote.IPurchaseDelegate;
 import games.strategy.triplea.delegate.remote.ITechDelegate;
-import games.strategy.triplea.oddsCalculator.ta.ConcurrentOddsCalculator;
-import games.strategy.triplea.oddsCalculator.ta.IOddsCalculator;
+import games.strategy.triplea.odds.calculator.ConcurrentOddsCalculator;
+import games.strategy.triplea.odds.calculator.IOddsCalculator;
 import games.strategy.triplea.ui.TripleAFrame;
 import games.strategy.util.CollectionUtils;
 import games.strategy.util.Tuple;
