+++ game-core/src/main/java/games/strategy/triplea/odds/calculator/OddsCalculatorDialog.java
@@ -1,4 +1,4 @@
-package games.strategy.triplea.oddsCalculator.ta;
+package games.strategy.triplea.odds.calculator;
 
 import java.awt.BorderLayout;
 import java.awt.Dimension;
