+++ game-core/src/main/java/games/strategy/triplea/oddsCalculator/ta/BattleResults.java
@@ -10,6 +10,8 @@ import games.strategy.triplea.delegate.IBattle.WhoWon;
 import games.strategy.triplea.delegate.Matches;
 import games.strategy.util.CollectionUtils;
 
+// TODO: Move to triplea.delegate or triplea.delegate.dataObjects package upon next incompatible release
+
 public class BattleResults extends GameDataComponent {
   private static final long serialVersionUID = 1381361441940258702L;
