+++ game-core/src/main/java/games/strategy/triplea/ui/menubar/GameMenu.java
@@ -29,7 +29,7 @@ import games.strategy.engine.framework.IGame;
 import games.strategy.engine.random.IRandomStats;
 import games.strategy.engine.random.RandomStatsDetails;
 import games.strategy.sound.SoundOptions;
-import games.strategy.triplea.oddsCalculator.ta.OddsCalculatorDialog;
+import games.strategy.triplea.odds.calculator.OddsCalculatorDialog;
 import games.strategy.triplea.settings.ClientSetting;
 import games.strategy.triplea.ui.PoliticalStateOverview;
 import games.strategy.triplea.ui.TripleAFrame;
